requirements.txt
requests~=2.22.0
python-dateutil~=2.8.2
pandas~=1.5.0
websocket-client~=1.4.1
ipdb~=0.13.9
cryptography~=38.0.1